infoprovider: add default GlueSE.ImplementationVersion
The GlueSE LDAP object has the attribute GlueSEImplementationVersion.  This value is 
calculated from the PoolManager cell.  If the info service doesn't have information about 
the PoolManager then the info-provider will attempt to publish the GlueSE object with an 
empty GlueSEImplementationVersion attribute.

Although publishing GlueSEImplementationVersion is declared optional in the LDAP Schema, if 
published it's not allowed to have an empty value.  An attempt to publish a GlueSE object 
with an empty GlueSEImplementationVersion results in OpenLDAP rejecting the whole GlueSE 
object.  A site publishing their dCache instance without a GlueSE object results in various 
clients failing with interesting errors.

The patch does two things: it adds a default hard-coded value ("UNDEFINEDVALUE") should the 
info-provider not have information about the PoolManager.  It also suppresses publishing 
GlueSEImplementationVersion should the value turn out to be empty (not-empty="1").

Either is sufficient to alieviate the problem; however, the WLCG Glue Schema usage document 
requires us to publish a GlueSEImplementation Version, hence the default value.

Acked-by: Gerd
RB: http://rb.dcache.org/r/1021/