infoprovider: support suppress elements inside attr elements
WLCG Glue Usage document describes how GlueSA VOInfo and GlueService objects all take
AccessControlBaseRule (ACBR) attributes with values of the form VO:vo-name (e.g.,
"VO:atlas").  For some GlueSA and VOInfo objects, these ACBR values are calculated from the
store units: we assume that a store unit maps to a single VO.

For some sites, it is not desirable to publish all store-units, if for no other reason that
they don't have an associated Grid VO.  To allow for this, the corresponding VO (in the
mapping from store-unit to VO name) is set to the empty string.  This results in publishing
ACBR values like "VO:".

I consider it a bug to publish these ACBR attributes with empty VOs.  Whilst I believe this
doesn't break anything, it doesn't follow the description of how ACBRs should be published.
It may also lead to undesirable results; e.g., if a client queries BDII and generates a
list of supported VOs it may interpret the empty VOs in some odd fashion.

Currently, Xylophone only supports suppressing of LDAP attributes if the attribute is
empty.  However, these attributes are never empty since they will always have the "VO:"
prefix.  Therefore the current support in Xylophone is insufficient to fix this problem.

This patch uses the existing support in Xylophone for predicates (used for dynamically
suppressing object publishing).  With this patch, the same predicate attribute ("suppress")
may be supplied within the attribute value.  If it evaluates true then the attribute is
suppressed.

Acked-by: Gerd
RB: http://rb.dcache.org/r/1027/