infoprovider: suppress publishing empty VO GlueSA.ACBR attributes
WLCG Glue Usage document describes how GlueSA VOInfo and GlueService objects all take 
AccessControlBaseRule (ACBR) attributes with values of the form VO:vo-name (e.g., 
"VO:atlas").  For some GlueSA and VOInfo objects, these ACBR values are calculated from the 
store units: we assume that a store unit maps to a single VO.

For some sites, it is not desirable to publish all store-units, if for no other reason that 
they don't have an associated Grid VO.  To allow for this, the corresponding VO (in the 
mapping from store-unit to VO name) is set to the empty string.  This results in publishing 
ACBR values like "VO:".

I consider it a bug to publish these ACBR attributes with empty VOs.  Whilst I believe this 
doesn't break anything, it doesn't follow the description of how ACBRs should be published.  
It may also lead to undesirable results; e.g., if a client queries BDII and generates a 
list of supported VOs it may interpret the empty VOs in some odd fashion.

This patch adds the suppress elements to prevent publishing ACBR values with empty VOs 
(values that are "VO:").  This patch is dependent on an earlier patch 
(http://rb.dcache.org/r/1027/)

Acked-by: Gerd
RB: http://rb.dcache.org/r/1028/