xrootd: Fix parse error for open request
I found the following error in our logs:

java.lang.StringIndexOutOfBoundsException: String index out of range: -4
        at java.lang.String.substring(String.java:1937)
        at org.dcache.xrootd2.protocol.messages.OpenRequest.getOpaqueMap(OpenRe$
        at org.dcache.xrootd2.door.XrootdRedirectHandler.doOnOpen(XrootdRedirec$
        at org.dcache.xrootd2.core.XrootdRequestHandler.messageReceived(XrootdR$
        at org.dcache.xrootd2.door.XrootdRedirectHandler.messageReceived(Xrootd$
        at org.jboss.netty.channel.SimpleChannelHandler.handleUpstream(SimpleCh$
        at org.jboss.netty.channel.DefaultChannelPipeline.sendUpstream(DefaultC$
        at org.jboss.netty.channel.DefaultChannelPipeline$DefaultChannelHandler$
        at org.jboss.netty.handler.execution.ChannelEventRunnable.run(ChannelEv$
        at java.util.concurrent.ThreadPoolExecutor$Worker.runTask(ThreadPoolExe$
        at java.util.concurrent.ThreadPoolExecutor$Worker.run(ThreadPoolExecuto$
        at java.lang.Thread.run(Thread.java:619)


I tracked this down to parsing opaque strings like '&s&s=s'. The flaw
was that the parser would not recognize that the equal sign it finds
when parsing the first token belongs to the second token.

Even with the fix, the request would fail as the token is not valid
opaque information. However now it fails with a proper error.

Target: 1.9.5
Target: trunk
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>
Patch: http://rb.dcache.org/r/1026/