info: fix summary of SRM reservations so changes in ownership propagate
Information about SRM reservations are published under the path: reservations.<id>.  This
information is primary data.

The ReservationByDescMaintainer maintains a summary of SRM reservations aggregating by (VO,
description), based on the SRM reservation primary data.  This derived data is published under the
path: summary.reservations.by-VO.<vo-name>.by-description.<description>.

The VO that the reservation was created for (the "owner") is part of how reservations are grouped
together.  However, the metric representing this was missing from the list of predicates that
trigger an update.  The effect is that, when a site-admin changes for whom the reservation was
created, the reservation information updates but the summary information is not.

The summary information is used to publish SRM tokens in GLUE, so this bug prevents any change to a
reservation's "ownership" to propagate to GLUE.  The information will eventually propagate when one
of the other predicates is triggered (e.g, used space changes).  Alternatively, the info service
could be restarted.

This patch adds the missing predicate.  It also adds a unit-test to verify that updating the
ownership has the desired effect.

Acked-by: Gerd
RB: http://rb.dcache.org/r/1035