Treat any exceptions thrown by getStorageInfo in the same fashion as verification failures
Perhaps not surprisingly, the getStorageInfo methods can throw exceptions.  This happens when the 
underlying (PNFS) storage is somehow in an inconsistent fashion.  Currently, the comparator dies 
when this happens without indicating where the problem lies.  This is very inconvenient to the 
site-admin.

Dmitry (at DESY) gave me a list of the exceptions he sees.  This patch treats these exceptions in 
the same fashion as a MismatchException: by default the application will terminate, logging which 
PNFS-ID caused the problem but the user has the option to continue, discovering all problems with 
a single pass (the "-k" option).

NB I know the code is ugly: I will be refactoring it soon.

Acked-by: Tigran (reluctantly)
RB: http://rb.dcache.org/r/1032/