infoprovider: fix typos found by Stephen Burke
Stephen Burke was generous enough to cast his expert eye over the output from NDGF.  He noticed a 
number of anomalies, some of which are simply typos in the info-provider.

This patch fixes those issues that are clearly typos.  The changes are:
  1. Fix capitalisation of AL/RP values ("Replica" instead of "REPLICA")
  2. Fix xroot protocol version (dCache protocol family seems to have changed from "Xrootd" to 
     "root")
  3. Change Capacity attribute values from Installed*Capability to Installed*Capacity
  4. Don't publish deprecated AccessProtocol.AccessTime attribute
  5. Don't publish deprecated AccessProtocol.Port attribute

My thanks to Stephen for his help here.

Acked-by: Gerd
RB: http://rb.dcache.org/r/1040/