PnfsManager: Call correct constructor in PnfsGetFileMetaDataMessage
PnfsGetFileMetaDataMessage inherits from PnfsGetFileAttributes. The two
classes have different interpretations of a string argument in the
constructor. PnfsGetFileMetaDataMessage interprets it as a PnfsId
whereas PnfsGetFileAttributes interprets it as a path. This lead to the
wrong constructor being called. The problem is resolved by this patch.

Target: trunk
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>
Patch: http://rb.dcache.org/r/1042/