jglobus: Add CRL caching
JGlobus recently added support for CRLs. In every session it checks
whether CRLs have been modified since they were loaded and reloads them,
if necessary. As the CRL in memory representation is a global shared
object, the check of the last modification time is performed under a
global lock. This turns out to be a bottleneck when performing many
concurrent handshakes (e.g. on the SRM server). This bottleneck prevents
the SRM from scaling.

The patch avoids the bottleneck by introducing a cache time. It prevents
the CRL files from being checked on every session, thus reducing the
bottleneck. Notice that the patch doesn't remove the requirement for
global synchronization. The patch does however reduce the frequency of
refresh operation to a point where it is no longer a bottleneck.

We have deployed this patch under our production server. Without this
patch, I could push our server to 140Hz SRM pings. With this patch, I
can push it to 295 Hz.

Target: trunk
Acked-by: Timur Perelmutov <timur@fnal.gov>
Patch: http://rb.dcache.org/r/1057/