chimera migration: refactor comparator, extract comparison logic
This patch is the first in a series of patches that aim to improve the overall performance of the
tool for checking the PNFS to Chimera migration.

This patch refactors the comparator by extracting the checking logic.  It introduces the concept of
a PNFS-ID Validator by adding the PnfsIdValidator Interface.  Classes that implement PnfsIdValidator
provide some indicator whether the file represented by the PnfsId is somehow OK.

The test logic currently in Comparator is split into three classes that implement PnfsIdValidator:
FileMetaDataComparator, StorageInfoComparator and CombinedComparator.

FileMetaDataComparator checks that the FileMetaData from two (arbitrary) NameSpaceProvider objects
is equal.  Should the FileMetaData differ, it can optionally emit information to a PrintStream.

The StorageInfoComparator provides a similar functionality to FileMetaDataComparator: it checks the
StorageInfo returned by two NameSpaceProvider objects and (optionally) emits information should they
differ.

The CombinedComparator provides a logical AND, but without short-circuit evaluation.  It can be
created with multiple PnfsIdValidator objects and will check each PnfsIdValidator.  The result is
only "OK" if all member PnfsIdValidator objects evaluate to "OK".  It avoids short-circuit to allow
all member PnfsIdValidators to check the PnfsId should the first member's isOK method return false.

The unit tests are also refactored and enhanced.

Acked-by: Gerd
RB: http://rb.dcache.org/r/1039/