Fix to include path for libdcap-tunnel-gsi-1.9.3-6.x86_64.rpm compilation against EPEL
The main change is the fix to the include path, bypassing the short cuts in the Make file.

The make file could be simplified, (since I am bias toward the EPEL distribution rather than VDT globus) but then the autotools port is coming and so this is a for now solution requiring minimal changes.

While doing this I also included removing some Obsoletes, since although all these packages do Obsolete the old name, since the dcap package provides as well as obsoletes, it can cause issues with

yum groupinstall glite-UI

when you want your rpms installed then do an upgrade. As some times the libs declare that they obsolete the dcap binary rpm. This is the sort of thing that should not happen and I assume its a multi arch yum bug as it does not occur with sl5 i386 but that could be because "groupinstall" is not set up.

Signed-off by: No one
http://rb.dcache.org/r/1033/