xylophone: use indirect reference to external XML file locations
This patch alters xylophone to use indirect reference for the location of an external XML file.

XSLT expects that a source delivers at least some contents; if this doesn't happen (e.g., a file 
doesn't exist) then a typical XSLT processor will emit a warning.

The decision during yesterday's dev-meeting was that site-admin generated content for the 
info-provider be stored in /var/opt/dcache (or /var/lib/dcache).  This is (deliberately) outside 
our deployment tree.  To avoid any warnings, we must provide an "empty" (but valid XML) file that 
is used by default.  There are two ways to do this: either we ensure that 
/var/opt/dcache/tape-info.xml exists within the info-provider script (copying from a dummy file as 
necessary) or the glue-1.3.xml refers to some "empty" file within /opt/d-cache by default but we 
make it very easy for the site admin to change the location to /var/opt/dcache/tape-info.xml.  I 
favour the latter of the two options.

As an aside: the XML format allows one to define entities; these are like macros that expand to 
their defined content.  It would be possible to use these to provide a common place to define the 
location of the tape-info.xml file.  However, this solution is rather ugly.

Instead, this patch introduces location definitions.  A location is defined within a locations 
element.  Such locations have a name attribute that may be used in a lookup, object repetition or 
sum-all operations.

Acked-by: Gerd
RB: http://rb.dcache.org/r/1069/