infoprovider: add support for nearline accounting
Sites have be requested (are required?) to publish near-line accounting information.  They are 
required to publish usage after taking compression into account, so dCache cannot provide these 
numbers.

To provide some support, we introduce a new file: tape-info.xml.  This contains information about 
the tape used in an HSM-independent fashion.  The anticipation is that site-admins provide 
information in this format (e.g., by writing a script that generates the file) and that the 
info-provider reads the file to ascertain what information should be published.

Acked-by: Gerd
RB: http://rb.dcache.org/r/1070/