add CellVersion support to AbstractCell
The cells package includes the concept of cells have a version.  This concept is encapsulated in 
the CellVersion class, which provides a version, release and revision strings, where version is 
always "<release>(<revision>)".

The CellInfo object includes the CellVersion object.  When asked for the CellInfo, the CellNucleus 
obtains CellVersion by calling (via reflection) getCellVersion method on the cell.  CellAdapter, 
which AbstractCell extends, contains a default getCellVersion method that returns some default 
values.

Within dCache, the LoginManager's getCellVersion() method returns information from the underlying 
Cell's getCellVersion.  However, since LoginManager contains the Class of the underlying cell, it 
needs a static method to obtain this information.  Therefore, all doors also include a 
getStaticCellVersion() static method as a hack around this problem.

The utility class diskCacheV111.util.Version provides the static method "getVersion" for obtaining 
the value of java.lang.Package.getSpecificationVersion from the manifest of whichever jar file 
contains diskCacheV111.util.  The cells in dCache use Version.getVersion to populate the release 
string in the CellVersion.

The problem is that AbstractCell does not include a getCellVersion method.  Because of this, the 
CellVersion inside CellInfo defaults to the value from CellAdapter, which simply says "cells" 
rather than the dCache version.  This problem affects all cells that extend CellAdapter, include 
UniversalSpringCell.

This is particularly a problem since the PoolManager now uses UniversalSpringCell, so the info 
service no longer reports the dCache version of the PoolManager (instead publishing just "cells").  
The info-provider publishes the PoolManager's dCache version as GlueSEImplementationVersion.  
Currently, this is published as "cells".

This patch adds the getCellVersion() method for the AbstractCell method.  It also includes the 
getStaticCellVersion so that a LoginManager will still provide correct information when used with 
an AbstractCell (or subclass thereof).

Acked-by: Gerd
RB: http://rb.dcache.org/r/1074/