xylophone: add support for unique constraints for objects and attributes
When publishing information about a group of pools in GLUE, we identify which VO corresponds to a
given store-unit.  However, a set of pools may be accessible from multiple store-units.  Some of
these store-units may map to the same VO.  Under these circumstances, we will currently publish
duplicate LDAP attribute in the GlueSA object; for example:

  [...]
  GlueSAAccessControlBaseRule: VO:atlas
  GlueSAAccessControlBaseRule: VO:atlas
  [...]

and, if the pools are writeable, multiple VOInfo objects.  For example:

  GlueVOLocalID: atlas:STATIC@osm
  GlueVOPath: /pnfs/example.org/data/atas
  GlueVOAccessControlBaseRule: VO:atlas

  GlueVOLocalID: atlas:LOCAL@osm
  GlueVOPath: /pnfs/example.org/data/atas
  GlueVOAccessControlBaseRule: VO:atlas

This patch adds support for a new attribute in xylophone: the unique attribute.  This is used to
describe uniqueness constraints that published objects or attributes must satisfy.  For each
object/attribute (from an iteration over objects in an external XML file) the "unique" element is
expanded.  If this expanded value matches the value from a previous object/attribute then this
object/attribute is not published.

With unique elements, it is possible to suppress duplicate information, so reducing the number of
object and attributes a site publishes.

Acked-by: Gerd
RB: http://rb.dcache.org/r/1080