gPlazma: Update client API to use CellStub
AuthzQueryHelper is a client API for gPlazma. The class had redundant
references to the CellEndpoint as it both requires the CellEndpoint in
the constructor and on every method.

This patch updates the AuthzQueryHelper to use CellStub in the
constructor. This elliminates the need to specify the endpoint and
CellPath on every method call. It also internally eliminates lots of
redundant message processing code and simplifies exception handling.

One issue here is that CellStub embodies our conventions for message
handling in dCache. As a result it can eliminate lots of repeated error
handling. It does however rely on those conventions being followed.
Unfortunately gPlazma does not follow those conventions, and hence the
client stub still needs some of the error handling that is otherwise
done by CellStub. We should try to clean this up later by revisiting the
message exchange between the client stub and the gPlazma cell.

One issue that I found while testing this patch is that delegation to
gPlazma cell was broken before this patch. The issue is that the
gPlazmaDelegationInfo class is not serializable. I have made it
serializable now. I do however wonder who uses this feature and for what
purpose: Given that we had not had any bug reports about this, I
speculate that not many have tried to use this.

I have updated the callers of AuthzQueryHelper to match the new
interface.

I updated AuthorizationException with new constructors to allow for a
cause to be specified.

Target: trunk
Acked-by: Timur Perelmutov <timur@fnal.gov>
Patch: http://rb.dcache.org/r/1071/