webdav: Add proxied write support
This patch adds write support to the WebDAV door. As I don't the exact
semantics of an HTTP PUT redirect and because Milton doesn't support it,
I chosen to use the door as a proxy for the first implementation. I.e.
the data will go from the client to the door and then from the door to
the pool.

To avoid adding a new mover, I use the existing GFtp mover in mode S.

I know the following things need to be done, but I would like to do
those in a separate patch:

- Update read path to use the new Transfer class
- Refactor the Transfer class hierarchy so it is not embedded into
  the DcacheResourceFactory class.

Target: trunk
Acked-by: Paul Millar <paul.millar@desy.de>
Patch: http://rb.dcache.org/r/1084/