jglobus: Upstream version of CRL cache patch
r12790 added a CRL cache patch to our jglobus fork. That patch was 
submitted to JGlobus and the JGlobus team found a minor typo in the 
patch. The typo caused the key pair cache timeout parameter to be used 
for definining the CRL cache timeout. The JGlobus team fixed the typo 
and sent back an updated version of our patch.

Target: trunk
Target: 1.9.5
Acked-by: Paul Millar <paul.millar@desy.de> (over jabber)