PoolManager: Fix the magic feature
PoolManager got a magic feature (yeah, that's actually what it's
called). When enabled (the default) it will internally update the cost
information of a pool whenever it selects a pool as a transfer
destination. This is essential in the current PoolManager as cost
updates from the pool are only generated every 30 seconds.

Unfortunately when refactoring the PoolManager ahead of 1.9.3 this
feature was broken: It relies on CostModuleV1 intercepting the replies
going back to the door upon pool selection. The message was
intercepted, but not dispatched internally.

Target: trunk
Target: 1.9.5
Target: 1.9.4
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>
Patch: http://rb.dcache.org/r/1105/