pool: Don't log a kill of a missing mover as an error
Some doors kill movers when the client disconnects. Sometimes the mover
has already terminated at that point and hence not finding the mover is
not an error. This patch reduces the log level of the message from
ERROR to INFO.

Target: trunk
Target: 1.9.5
Target: 1.9.4
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>
Patch: http://rb.dcache.org/r/1107/