PnfsManager: Refactor getFileMetaData
We are in the process of getting rid of the FileMetaDataClass.

Before getFileAttributes would use getFileMetaData internally. If we
don't need all those attributes, this wastes a lot of time fetching
information we will throw away. This patch refactors
BasicNameSpaceProvider such that getFileAttributes can fetch the
information that it would normally get through getFileMetaData directly.
The getFileMetaData method is updated to call getFileAttributes
internally and convert the resulting FileAttributes object to a
FileMetaData object.

Reading of .(getattr) is moved to PnfsFile and the
PnfsFile.getFileMetaData method is removed. PnfsFile.exists is updated
accordingly. PnfsFileMetaDataSource is removed as it has been deprecated
in a previous release.

Target: trunk
Acked-by: Paul Millar <paul.millar@desy.de>
Patch: http://rb.dcache.org/r/1112/