general: fix compiler warnings about UTF-8 literals
When compiling dCache, the compiler emits some warnings about some files containing unmappable files.
These characters are UTF-8 byte-sequences injected directory into String literal values.  This is the
wrong way to introduce these values. Java uses UTF-16 internally and supports a markup to allow the
inclusion of Unicode characters: the sequence "\uNNNN" represents the Unicode character NNNN.

This patch fixes this by replacing the funny k-like character with a "k" in the word "kill" (it looks
like a typo) and provides the proper escaping of the +/- character in the two places it is used.

Acked-by: Gerd
RB: http://rb.dcache.org/r/1115/