fixed RepositorySubsystemTest and StorageInfoTest
fixed above unittests.
For the RepositorySubsystemTest the sleepduration was not correctly related to the one in CacheRepositoryV5. Changed that.
For the StorageInfoTest the equals method was changed for migration purposes and the unittest was left out. Corrected that as talked to Tigran and Paul.
Target:
trunk Committed revision 12925.
1.9.5 

http://rb.dcache.org/r/1118/

acked by Gerd