PnfsManager: Add SIMPLE_TYPE and optimize listing
To maximize list performance, we introduce SIMPLE_TYPE as a simpler
version of the TYPE FileAttribute. SIMPLE_TYPE is not guaranteed to (but
may) identify symlinks and thus can avoid resolving PNFS ID and reading
the .(getattr) file.

Listing is refined such that for SIMPLE_TYPE, SIZE, and
MODIFICATION_TYPE we avoid the call to the getFileAttributes method.

Target: trunk
Acked-by: Dmitry Litvintsev <litvinse@fnal.gov>
Patch: http://rb.dcache.org/r/1158/