Fix auto-discover of PNFS mount-points
When the PNFS namespace provider (BasicNameSpaceProvider) starts up, it needs to know the mount-point of 
PNFS on the local machine.  The user may supply this as a configuration option, but there is also an 
auto-detection mechanism.  I guess most people use the autodetection.

The auto-detection works by scanning a file that lists all current mount-points.  The location and 
format of this file is platform-specific.  On Linux machines it scans the /etc/mtab file.  For each 
mount-point listed in the file, the auto-detection code attempts some PNFS-specific queries using 
dot-commands.  If these commands succeed then the mount-point becomes a candidate for selection.  Once 
all the mount-points are tested a further refinement process reduces the list of candidate mount-points 
to (at most) a single mount-point, which is then used.

The problem is that, the attempt to access the dot-command file may result in IllegalArgumentException 
being thrown.  With IOException (which may also be thrown) the code catches the exception, emits a log 
message and moves on to the next listed mount-point.  However, IllegalArgumentException is not caught; 
the effect of this exception being thrown is to terminate the whole auto-detection process and no PNFS 
mount-points are discovered.

The patch fixes this by catching IllegalArgumentException and, as with IOException, rejecting the 
mount-point and moving onto the next potential PNFS mount-point.

RB: http://rb.dcache.org/r/1126/
Acked-by: Tigran
Target: 1.9.5, Trunk