update info-provider docs (README-GLUE) with fault-finding information
A recent query on user-forum asked (in effect) how to fault-find problem with the
info-provider where BDII is reporting failures.  The process is somewhat awkward and current
not well documented so the email reply was rather long.

This patch takes the information in the email, reformats it slightly and adds it as another
section at the end of the GLUE README file.  The hope is that this will enable site-admins
to diagnose some BDII problems themselves.

Acked-by: Gerd
RB: http://rb.dcache.org/r/1165/

Patch is for all supported branches (1.9.1, 1.9.2, 1.9.3, 1.9.4, 1.9.5 and Trunk)