infoprovider: add some require GLUE attributes that were missing
After an extended debugging session with Steve Traylen, there are a couple of attributes that the current
info-provider doesn't publish.  These are, for the SE object:

  GlueSESizeTotal,
  GlueSESizeFree

and for the SA objects representing SRM reservations:

  GlueSAReservedOnlineSize

This patch adds support for publishing these values.

The patch is for 1.9.5 and Trunk.

Acked-by: Gerd
RB: http://rb.dcache.org/r/1177/