infoprovider: publish namespace implementation within GlueSEImplementationVersion
GLUE v1.3 provides few options for publishing SE-wide additional information.  Logically, such
information belongs to the StorageElement object (LDAP GlueSE); however, this class provides little
opportunity for implementation-specifc additional user information.

The only realistic option is to use the ImplementationVersion property (the
GlueSEImplementationVersion LDAP attribute).  The 1.3 spec. says that this is "[t]he version of the
implementation" and the "Usage of Glue Schema v1.3 for WLCG Installed Capacity information" document
describes this property, saying "[GlueSEImplementationVersion] MUST be published. The format is
defined by the implementations."

I propose the following definition of the dCache ImplementationVersion property:

  {ARCHITECTURE} '.' {MAJOR} '.' {MINOR} '-' {PATCH} {RELEASE-METADATA}
  ' ' '(' {KEY-VALUE} [';' {KEY-VALUE}...] ')'

Where {KEY-VALUE} has the form:

  {KEY} '=' {VALUE}

The following are all examples that match this format:

  "1.9.4-1"
  "1.9.5-9"
  "1.9.5-10 (ns=Chimera)"
  "1.9.5-10 (ns=PNFS;foo=Bar)"
  "1.9.6-1rc (ns=Chimera;foo=Bar;baz=QUX)"

The disadvantage of this approach is that simple GLUE clients that show statistics based on aggregated
numbers over dCache instances with the same GlueSEImplementationVersion will no longer aggregate over
only the dCache version; for example, if the client wishes to show the number of SEs with "dCache"
version "1.9.5-10" then, after this patch, such a client will show "1.9.5-10 (ns=PNFS)" and "1.9.5-10
(ns=Chimera)" as different dCache versions.  This might or might not be desirable.

Patch is for 1.9.5 and trunk.

Acked-by: Gerd
RB: http://rb.dcache.org/r/1181/