Specify paths to JDBC files that contains the DDL statements used by toplink
This is a fix to problem reported in http://www.dcache.org/rt/Ticket/Display.html?id=4655
this patch specifies the paths to the JDBC files that contains the DDL statements used by toplink by toplink.drop-ddl-jdbc-file-name and toplink.create-ddl-jdbc-file-name options. I tried specifying absolute paths, but  this did not work, as it appears that toplink is considering all paths to be relevant to the CWD. So I changed the path to be relevant and added the change of the CWD in the srm startup script. 
I chose the directory /opt/d-cache/etc/ for storing reateDDL.jdbc and dropDDL.jdbc, because this is where we store other sql files.
In the future we might consider making this files part of the rpm, but for now they are created automatically the first time srm runs.

Signed-off by: Paul, Vijay
http://rb.dcache.org/r/1143/