cells: Fix and clean up admin shell
This patch restructurs StreamObjectCell, the admin shell used by dCache.
Cell initialization, socket handling, etc. is restructured to eliminate
as many fields as possible. The JLine ConsoleReader has received a more
significant role as the InputHandler method has been eliminated. That
means we can use more of the features of JLine such as prompt handling
and saving the history to a file.

One insignificant semantic change is that an ASCII session can now be
promoted to a binary session and any point we receive $BINARY$. Before
this could only happen at the beginning of a session.

The patch also fixes the problem described in ticket #5234. The patch is
too large to go into 1.9.5, so I suggest to write a separate patch to
fix the problem. In short the problem was that we closed the input and
output streams of the socket when logoff is issued. This immediately
closes the socket. Instead one should shut down the output of the socket
by calling shutdownOutput. Then we should wait for the client to hangup
and first then should we close the socket.

The placement of the history file is now configurable and defaults to
/var/opt/d-cache/adminshell_history.

Target: trunk
Ticket: #5234
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>
Patch: http://rb.dcache.org/r/1179/