scripts: Add commands to import certificates
Java libraries rely on KeyManagers and TrustManagers to access key and
certificate material. This is true for Jetty and thus for the new WebDAV
door, as well as for ActiveMQ and other libraries.

Ideally we would have a custom security provider that provides a
KeyManager and TrustManager for /etc/grid-security directly. gLite does
contain a KeyManager and a TrustManager for /etc/grid-security, however
they don't seem to be packaged as a security provider. Hence they are
not easy to instantiate in existing products. Eventually we should try
to repackage the gLite classes, but for now we have to resort to
converting the certificates in /etc/grid-security to PKCS12 for the host
certificate and a Java keystore file the CA certificates.

To make the conversion easy, this patch adds the 'dcache import
hostcert' and 'dcache import cacerts' commands. I had to fix a couple of
bugs in config.sed to make this work, as well as refactor some of the
support scripts.

The WebDAV configuration parameters are refined such that we can have
shared parameters for defining the location of the files.

Target: trunk
Acked-by: Paul Millar <paul.millar@desy.de>
Patch: http://rb.dcache.org/r/1164/