migration comparator: various improvements
The current comparator combines the logic to read Pnfs-IDs from a file and the logic that checks a stream
of Pnfs-IDs.  It also provides the end-user with a poor array of options to control which Chimera
configuration file is used and how the PNFS mount-point is obtained.

This patch factors out the logic for reading PnfsIDs from a file, which is made slightly more robust and
now reports problems (any malformed PNFS-ID values) along with the line-number in the file where the
problem occurred.

The patch introduces a blocking queue that is used to decouple the acquisition of PNFS-IDs from the
analysis of them.  This was originally done as a first-step towards increasing through-put but, due to
PNFS dbserver single-threaded nature, this will be harder to achieve.

The patch also introduces various comnmand-line options, so the location of the Chimera XML configuration
file is configurable, as is the PNFS mount point.  By default, both take their previous values
(/opt/d-cache/config/chimera-config.xml and auto-detect, respectively).

Another option allows the user to increase the logging output.  This may be useful in diagnosing any
problems with the PNFS mount-point auto-discovery process.

The patch is for 1.9.5 and trunk

Acked-by: Gerd
RB: http://rb.dcache.org/r/1125/