PoolManager: Extend magic feature to take file size into account
CostModule has a magic feature to adjust cpu cost when it selects a pool
for a particular operation. It did however not account for the changes
in free space. This is a problem when space cost factor and cpu cost
factor are zero. In that case an almost full pool is selected as often
as an empty pool. Thus it is quite possible that an almost full pool is
selected for several transfers and the sum of those files doesn't fit on
the pool. This problem is not the same problem we fixed last year: That
problem was that we would select pools if when they are already full.
The present bug is that the pool is not full yet, but the transfers
assigned don't all fit. This is not a problem with non-zero cost
factors, because an almost full pool will not be selected as often.

To fix this problem, the patch extends the magic feature to also adjust
the amount of "pinned space" (don't blame me for the terminology; that's
what it was called in PoolSpaceInfo). When we assign a write to a pool,
we increase the amount of pinned space and thus reduce the amount of
free space.

The WriteRequestHandler had to be updated to take all possible source of
a file size estimate into account.

The patch itself is not enough: As the adjustment is only effective
until the next pool up message, we need to fix the pool to prereserve
the total file size up front if the file size is already known.

We also need to fix the GridFTP door such that the size from the ALLO
command is used during pool selection. That size doesn't equal the file
size, but it is at least as large as the file. That's perfect for pool
selection and for preserving space in the pool. We need to add a similar
field to the PoolAcceptFileMessage.

Target: trunk
Target: 1.9.5
Ticket: #5075
Patch: http://rb.dcache.org/r/1171/
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>