Added command line options to process information system xml
The patch from ReviewBoard entry http://rb.dcache.org/r/1153/ was applied to Trunk as commit r12976.

This patch has not been applied to 1.9.5 branch, despite the ReviewBoard entry being marked "Branch: trunk, 
1.9.5". During the review process, it was my impression that Owen intended to commit this patch to the 1.9.5; 
however, due to illness, he is currently unavailable to do this himself.

The patch associated with this request is identical to that committed to Trunk as r12976.  Since I reviewed the 
original patch, I'm asking someone else to ACK this.

Patch is for 1.9.5

Acked-by: Gerd
RB: http://rb.dcache.org/r/1195/