Make dCacheConfigure.sh work with new information system configuration management ruby script.
The patch from ReviewBoard entry http://rb.dcache.org/r/1157/ was applied to Trunk as commit r12979.  This patch 
has not been applied to the 1.9.5 branch.

There is nothing in the above ReviewBoard to indicate that that patch was indented for the 1.9.5; however, 
during the review process, it was my impression that Owen intended to commit this patch to the 1.9.5.  Due to 
illness, he is currently unavailable to do this himself.

In the interests of expediency, I'm submitting the identical patch that Owen committed to Trunk as r12979.  
Since I reviewed the original patch, I'm asking someone else to ACK this.

Please note that this patch depends on the functionality supplied by http://rb.dcache.org/r/1195/

Patch is for 1.9.5 only

Acked-by: Gerd
RB: http://rb.dcache.org/r/1196/