comparator: add script to run comparator
The dCache code-base includes a PNFS-to-Chimera comparator: a stand-alone application that uses existing dCache
code to check that, for a list of file PNFS-IDs, PNFS and Chimera name-space providers will supply information
that is equal.

For historical reasons, the script that invokes this application was packaged separately.  However, this now
makes little sense: it increases the effort required to conduct the test without any benefit.

This patch provides an initial version of the script used to invoke the comparator.

The patch is for 1.9.5 and trunk.

Acked-by: Gerd
RB: http://rb.dcache.org/r/1199/