WebDAV: Refactor Transfer class; add billing and transfer monitoring
This patch moves the Transfer class out of the DcacheResourceFactory
class. The Transfer class is independent of WebDAV transfers and could
be reused in other doors that perform blocking transfers.

Subclasses of the Transfer class are still embedded in
DcacheResourceFactory, as they need access to the field in
DcacheResourceFactory.

A Transfer instance is now used for both uploads and downloads, and a
consequence is that we now have billing and transfer monitoring for
both.

Emulation of LoginManager has been completed. This is essential for
transfer monitoring to work.

Target: trunk
Acked-by: Paul Millar <paul.millar@desy.de>
Patch: http://rb.dcache.org/r/1130/