srm: PutCompanion: do not try to unregister dir creators if dir creation was not attempted.
If PutCompanion receives exception or timeout, it calls unregisterAndFailCreator() which performs the cleanup and failure of the requests that are writing in the same path and wait for the creation of the same directories. If the exception or timeout happens before the dir creation was attempted, the function call fails with the exception. This patch addresses this by checking that current_dir_depth != -1 (default value which would indicate that directory creation was not attempted) before we call unregisterAndFailCreator. 
 
Signed-off by: Vijay
http://rb.dcache.org/r/1197/