infoprovider: add support for default store unit (*@*)
The *@* store unit has a complex behaviour: it matches all requests that have no explicit store
unit in the namespace, so acts as a default store unit.  In general, which VOs' members may make
requests that attract the *@* store unit depends on the namespace permissions, mapping from VO to
namespace (user,group) and where (in the namespace) the store units are defined.

The current info provider makes the assumption that each store unit is for a well-defined, single
VO.  This is used to map store-units to their corresponding VOs when publishing accessible space
(the SA objects in Glue).  Implicit in this assumption is that *@* is not used to select pools for
more than one VO.

For large sites the assumption will likely hold: those sites may have a store unit per VO and there
are likely to be no VOs that access data through the *@* store unit.  For smaller sites this is not
the case: they may use the default store unit to select pools.

This patch addresses the lack of support for *@*.  It removes this special store unit from the
normal VO-based publishing.  It also allows the user to specify a list of VOs that may use *@*
default store unit and LDAP components (attributes and objects) are published to include that VOs
may access pools using the *@* store unit.

The patch is for 1.9.5 and trunk

Acked-by: Gerd
RB: http://rb.dcache.org/r/1208/